display: edo panel: display quality optimization
Because edo panel design using delta pixel design,
so you can see aliasing in the display edges
of the panel. In order to optimize the display
quality,add algorithms to rending the framebuff data.

If we detect the display pixel design is Delta,
disable mdpcomp to get the whole single layer framebuff,
then use the algorithms to optimize the pixel arrangement.

Change-Id: I4d6ac7e2bae13548ed35807eb6a2b7d933101cb4
Signed-off-by: l00212897 <gang.li@huawei.com>